src/requirements.txt
streamlit
langchain
openai
python-dotenv
google-search-results
faiss-cpu
requests
wikipedia-api
youtube-search-python

